+++ cli/cli/Commands/Content/ContentSnapshotCommand.cs
@@ -1,4 +1,5 @@
 ﻿using Beamable.Common.BeamCli;
+using Beamable.Common.BeamCli.Contracts;
 using System.CommandLine;
 
 namespace cli.Content;
@@ -15,11 +16,11 @@ public class ContentSnapshotCommand : AtomicCommand<ContentSnapshotCommandArgs, 
 	{
 		AddOption(new Option<string>("--manifest-id", () => "global", "Defines the name of the manifest that the snapshot will be created from. The default value is `global`"), (args, s) => args.ManifestIdToSnapshot = s);
 		AddOption(new Option<string>("--name", () => "", "Defines the name for the snapshot to be created"), (args, s) => args.SnapshotName = s, new[] { "-n" });
-		AddOption(new Option<ContentSnapshotType>("--snapshot-type", () => ContentSnapshotType.Local,
+		AddOption(new Option<string>("--snapshot-type", () => nameof(ContentSnapshotType.Local),
 				"Defines where the snapshot will be stored to." +
 				$"\n{nameof(ContentSnapshotType.Local)} => Will save the snapshot under `.beamable/temp/content-snapshots` folder" +
 				$"\n{nameof(ContentSnapshotType.Shared)} => Will save the snapshot under `.beamable/content-snapshots` folder"),
-			(args, type) => args.ContentSnapshotType = type, new[] { "-t" });
+			(args, type) => args.ContentSnapshotType = Enum.Parse<ContentSnapshotType>(type), new[] { "-t" });
 	}
 
 	public override async Task<ContentSnapshotResult> GetResult(ContentSnapshotCommandArgs args)
@@ -35,6 +36,7 @@ public class ContentSnapshotCommand : AtomicCommand<ContentSnapshotCommandArgs, 
 	}
 }
 
+[CliContractType, Serializable]
 public class ContentSnapshotCommandArgs : ContentCommandArgs
 {
 	public string ManifestIdToSnapshot;
@@ -46,11 +48,4 @@ public class ContentSnapshotCommandArgs : ContentCommandArgs
 public class ContentSnapshotResult
 {
 	public string SnapshotFullPath;
-}
-
-[CliContractType, Serializable]
-public enum ContentSnapshotType
-{
-	Local,
-	Shared
 }