+++ cli/cli/Commands/Content/ContentSnapshotListCommand.cs
@@ -1,6 +1,9 @@
 ﻿using Beamable.Common.BeamCli;
+using Beamable.Common.BeamCli.Contracts;
+using Beamable.Common.Docs;
 using Beamable.Server;
 using System.Collections.Concurrent;
+using System.CommandLine;
 using System.Text.Json;
 
 namespace cli.Content;
@@ -15,6 +18,7 @@ public class ContentSnapshotListCommand : AtomicCommand<ContentSnapshotListComma
 
 	public override void Configure()
 	{
+		AddOption(new Option<string>("--manifest-id", () => "global", "Defines the name of the manifest that will be used to compare the changes between the manifest and the snapshot. The default value is `global`"), (args, s) => args.ManifestId = s);
 	}
 
 	public override async Task<ContentSnapshotListResult> GetResult(ContentSnapshotListCommandArgs args)
@@ -23,6 +27,9 @@ public class ContentSnapshotListCommand : AtomicCommand<ContentSnapshotListComma
 		
 		var localSnapshotPaths = _contentService.GetContentSnapshots(true);
 		var sharedSnapshotPaths = _contentService.GetContentSnapshots(false);
+
+		var allContentFiles = await _contentService.GetAllContentFiles(manifestId:args.ManifestId);
+		
 
 		var parseLocalSnapshots = localSnapshotPaths.Select(async path => await PaseManifestSnapshot(path));
 		var parseSharedSnapshots = sharedSnapshotPaths.Select(async path => await PaseManifestSnapshot(path));
@@ -47,15 +54,49 @@ public class ContentSnapshotListCommand : AtomicCommand<ContentSnapshotListComma
 			{
 				string manifestContent = await File.ReadAllTextAsync(path);
 				var manifestSnapshot = JsonSerializer.Deserialize<ManifestSnapshot>(manifestContent, ContentService.GetContentFileSerializationOptions());
-				var contentSnapshotListItems = manifestSnapshot.ContentFiles.Select(contentSnapshot => new ContentSnapshotListItem()
+				var contentSnapshotListItems = manifestSnapshot.ContentFiles.Select(contentSnapshot =>
 				{
-					Name = contentSnapshot.Key, Checksum = contentSnapshot.Value.Checksum,
-				}).ToArray();
+					// Lets check what the current status of the content to be restored
+					ContentStatus status = ContentStatus.Invalid;
+					int contentManifestReferenceIndex = allContentFiles.ContentFiles.FindIndex(content => content.Id == contentSnapshot.Key);
+					// If the content doesn't exist on manifest, it means that it is new
+					if (contentManifestReferenceIndex == -1)
+					{
+						status = ContentStatus.Created;
+					}
+					else
+					{
+						var contentManifestRef = allContentFiles.ContentFiles[contentManifestReferenceIndex];
+						switch (contentManifestRef.GetStatus())
+						{
+							case ContentStatus.Modified:
+							case ContentStatus.UpToDate:
+							case ContentStatus.Created:
+								status = contentManifestRef.PropertiesChecksum == contentSnapshot.Value.Checksum
+									? ContentStatus.UpToDate
+									: ContentStatus.Modified;
+								break;
+							case ContentStatus.Deleted:
+								status = ContentStatus.Created;
+								break;
+						}
+					}
+					return new ContentSnapshotListItem() { Name = contentSnapshot.Key, CurrentStatus = (int) status, };
+				}).ToList();
+				
+				// All contents that doesn't exist in the snapshot will be deleted
+				contentSnapshotListItems.AddRange(allContentFiles.ContentFiles
+					.Where(content => !manifestSnapshot.ContentFiles.ContainsKey(content.Id) && content.GetStatus() != ContentStatus.Deleted).Select(content =>
+						new ContentSnapshotListItem() { Name = content.Id, CurrentStatus = (int)ContentStatus.Deleted }));
+				
 				return new ManifestSnapshotItem()
 				{
 					Name = Path.GetFileNameWithoutExtension(path),
 					Path = path,
-					Contents = contentSnapshotListItems
+					Pid = manifestSnapshot.Pid,
+					ManifestId = manifestSnapshot.ManifestId,
+					SavedTimestamp = manifestSnapshot.SnapshotTimestamp,
+					Contents = contentSnapshotListItems.ToArray()
 				};
 			}
 			catch (Exception e)
@@ -70,6 +111,7 @@ public class ContentSnapshotListCommand : AtomicCommand<ContentSnapshotListComma
 
 public class ContentSnapshotListCommandArgs : ContentCommandArgs
 {
+	public string ManifestId;
 }
 
 [CliContractType, Serializable]
@@ -84,13 +126,16 @@ public class ManifestSnapshotItem
 {
 	public string Name;
 	public string Path;
+	public string ManifestId;
+	public string Pid;
+	public long SavedTimestamp;
 	public ContentSnapshotListItem[] Contents;
-	public ContentSnapshotType Type;
 }
 
 [CliContractType, Serializable]
 public class ContentSnapshotListItem
 {
 	public string Name;
-	public string Checksum;
+	public int CurrentStatus;
+	public ContentStatus StatusEnum => (ContentStatus)CurrentStatus;
 }